refactor: move scenario related models
