+++ packages/app-builder/src/components/CaseManager/DecisionPanel/DecisionPanel.tsx
@@ -20,7 +20,7 @@ import { ScoreModifier } from '@app-builder/components/Scenario/Rules/ScoreModif
 import useIntersection from '@app-builder/hooks/useIntersection';
 import { type Pivot, type TableModel } from '@app-builder/models';
 import { type Decision, type RuleExecution } from '@app-builder/models/decision';
-import { type ScenarioIterationRule } from '@app-builder/models/scenario-iteration-rule';
+import { type ScenarioIterationRule } from '@app-builder/models/scenario/iteration-rule';
 import { type loader } from '@app-builder/routes/_builder+/cases+/$caseId+/_index';
 import { Await, useLoaderData } from '@remix-run/react';
 import clsx from 'clsx';
