+++ packages/app-builder/src/components/Scenario/TestRun/TestRunDetails.tsx
@@ -1,5 +1,5 @@
 import { type User } from '@app-builder/models';
-import { type ScenarioIterationWithType } from '@app-builder/models/scenario-iteration';
+import { type ScenarioIterationWithType } from '@app-builder/models/scenario/iteration';
 import { type TestRun } from '@app-builder/models/testrun';
 import { useTranslation } from 'react-i18next';
 import { Avatar } from 'ui-design-system';
