+++ packages/app-builder/src/models/index.ts
@@ -3,5 +3,5 @@ export * from './auth-errors';
 export * from './data-model';
 export * from './http-errors';
 export * from './marble-session';
-export * from './scenario-validation';
+export * from './scenario/validation';
 export * from './user';
