+++ packages/app-builder/src/routes/_builder+/scenarios+/$scenarioId+/test-run+/index.tsx
@@ -10,7 +10,7 @@ import {
 import { testRunsFilterNames } from '@app-builder/components/Scenario/TestRun/Filters/filters';
 import { TestRunSelector } from '@app-builder/components/Scenario/TestRun/TestRunSelector';
 import { isForbiddenHttpError, isNotFoundHttpError, type User } from '@app-builder/models';
-import { adaptScenarioIterationWithType } from '@app-builder/models/scenario-iteration';
+import { adaptScenarioIterationWithType } from '@app-builder/models/scenario/iteration';
 import { CreateTestRun } from '@app-builder/routes/ressources+/scenarios+/$scenarioId+/testrun+/create';
 import { initServerServices } from '@app-builder/services/init.server';
 import { useOrganizationUsers } from '@app-builder/services/organization/organization-users';
